M5nr.getSequences.rb new option -r

+++ Scripts/M5nr.getSequences.rb
@@ -12,7 +12,7 @@ require 'restclient'
 require 'open-uri'
 require 'JSON'
 
-o = {:q=>FALSE, :url=>'http://api.metagenomics.anl.gov/m5nr', :max=>0}
+o = {:q=>FALSE, :url=>'http://api.metagenomics.anl.gov/m5nr', :max=>0, :recover=>FALSE}
 OptionParser.new do |opts|
    opts.banner = "
 Downloads a set of sequences from M5nr with a given functional annotation.
@@ -24,6 +24,7 @@ Usage: #{$0} [options]"
    opts.separator ""
    opts.separator "Options"
    opts.on("-m", "--max INT", "Maximum number of sequences to download.  By default: all (0)."){ |v| o[:max] = v.to_i }
+   opts.on("-r", "--recover", "If set, tries to recover a previous FastA."){ o[:recover]=TRUE }
    opts.on("-n", "--url STR", "URL for M5nr API.  By default: #{o[:url]}."){ |v| o[:url] = v }
    opts.on("-o", "--out FILE", "File containing the sequences.  By default: value of -f appended with .fa."){ |v| o[:out] = v }
    opts.on("-q", "--quiet", "Run quietly (no STDERR output)"){ o[:q] = TRUE }
@@ -37,8 +38,21 @@ abort "-f is mandatory" if o[:function].nil?
 o[:out] = "#{o[:function].gsub(/ /,'_')}.fa" if o[:out].nil?
 uri_fun = URI::encode(o[:function])
 
+ignore = {}
+if o[:recover] and File.exists? o[:out]
+   ih = File.open(o[:out], "r")
+   ih.each_line do |ln|
+      id = /^>(\S+)\s/.match(ln)
+      unless id.nil?
+         ignore[id[1]] = 1
+      end
+   end
+   ih.close
+   of = File.open(o[:out], "a+")
+else
+   of = File.open(o[:out], "w")
+end
 
-of = File.open(o[:out], "w")
 next_url = "#{o[:url]}/function/#{uri_fun}"
 i = 0
 loop do
@@ -47,11 +61,13 @@ loop do
    abort "Unable to reach MG-RAST M5nr API, error code #{res_fun.code}." unless res_fun.code == 200
    fun = JSON.parse(res_fun.to_str)
    fun["data"].each do |datum|
-      res_seq = RestClient.get "#{o[:url]}/md5/#{datum["md5"]}", {:params=>{:sequence=>1}}
-      abort "Unable to reach MG-RAST M5nr API, error code #{res_seq.code}." unless res_seq.code == 200
-      seq = JSON.parse(res_seq.to_str)
-      of.puts ">#{datum["source"]}:#{datum["accession"]} #{datum["function"]} [#{datum["organism"]} taxid:#{datum["ncbi_tax_id"]}]"
-      of.puts seq["data"]["sequence"].scan(/.{80}|.+/).map{ |x| x.strip }.join($/)
+      if ignore["#{datum["source"]}:#{datum["accession"]}"].nil?
+	 res_seq = RestClient.get "#{o[:url]}/md5/#{datum["md5"]}", {:params=>{:sequence=>1}}
+	 abort "Unable to reach MG-RAST M5nr API, error code #{res_seq.code}." unless res_seq.code == 200
+	 seq = JSON.parse(res_seq.to_str)
+	 of.puts ">#{datum["source"]}:#{datum["accession"]} #{datum["function"]} [#{datum["organism"]} taxid:#{datum["ncbi_tax_id"]}]"
+	 of.puts seq["data"]["sequence"].scan(/.{80}|.+/).map{ |x| x.strip }.join($/)
+      end
       i += 1
       break if o[:max]>0 and i >= o[:max]
    end # |datum|
